fix: alteracao tipo de campo

+++ src/main/java/school/sptech/cursos/model/Curso.java
@@ -28,7 +28,7 @@ public class Curso {
     @JoinColumn(name = "professor_id", referencedColumnName = "id")
     private Professor professor;
 
-    private BigDecimal porcentagem_lucro;
+    private Integer porcentagem_lucro;
 
     @OneToMany(mappedBy = "curso")
     private List<AvaliacaoCurso> avaliacoes;
